--- controle_estoque/Crud/config.ini
@@ -1,6 +1,0 @@
-[DEFAULT]
-dbhost = 127.0.0.1
-dbname = sistemadeva
-dbuser = andre
-dbpassword = rsp
-
