+++ controle_estoque/UI_Files/mainConfig.ui
@@ -102,7 +102,7 @@ border-bottom: 2px solid #069;</string>
     <property name="frameShadow">
      <enum>QFrame::Raised</enum>
     </property>
-    <widget class="QPushButton" name="bt_MovCaixa">
+    <widget class="QPushButton" name="bt_confEmpresa">
      <property name="geometry">
       <rect>
        <x>5</x>
@@ -167,7 +167,7 @@ margin-top: 0;
       <bool>true</bool>
      </property>
     </widget>
-    <widget class="QPushButton" name="bt_AReceber">
+    <widget class="QPushButton" name="bt_confUser">
      <property name="geometry">
       <rect>
        <x>185</x>
@@ -232,7 +232,7 @@ margin-top: 0;
       <bool>true</bool>
      </property>
     </widget>
-    <widget class="QPushButton" name="bt_APagar">
+    <widget class="QPushButton" name="bt_confFinanceiro">
      <property name="enabled">
       <bool>true</bool>
      </property>
@@ -300,7 +300,7 @@ margin-top: 0;
       <bool>true</bool>
      </property>
     </widget>
-    <widget class="QPushButton" name="bt_APagar_2">
+    <widget class="QPushButton" name="bt_confDB">
      <property name="enabled">
       <bool>true</bool>
      </property>
